requirements.txt
streamlit==0.82.0
google_api_python_client==2.42.0
protobuf==3.19.4
